(feat) change file from main to lib, changed main from prints to tests
