+++ Cargo.toml
@@ -4,3 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+
+[[bin]]
+name = "lib"
+path = "binary_search_insertion_index/src/lib.rs"
